requirements.txt
# ====================================
# FRAMEWORK WEB
# ====================================
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.6

# ====================================
# BASE DE DATOS
# ====================================
sqlalchemy==2.0.45
pymysql==1.1.2
cryptography==42.0.0  # Requerido por PyMySQL

# ====================================
# VALIDACIÓN DE DATOS
# ====================================
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.1.0
email-validator==2.1.0
annotated-types==0.7.0

# ====================================
# AUTENTICACIÓN Y SEGURIDAD
# ====================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ====================================
# VARIABLES DE ENTORNO
# ====================================
python-dotenv==1.2.1

# ====================================
# UTILIDADES
# ====================================
python-dateutil==2.8.2

# ====================================
# DEPENDENCIAS BASE
# ====================================
anyio==4.12.0
click==8.3.1
colorama==0.4.6
greenlet==3.3.0
h11==0.16.0
idna==3.11
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# ====================================
# DESARROLLO (Opcional - Descomentar si las necesitas)
# ====================================
# pytest==7.4.4
# pytest-asyncio==0.23.3
# httpx==0.26.0
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0
